requirements.txt
boto3>=1.34.0
botocore>=1.34.0
awsiotsdk>=1.21.0
python-dotenv>=1.0.1
